+++ tests/test_debug.py
@@ -1,24 +1,14 @@
 import pytest
 import pytest_benchmark
 
-pytest.skip("skipping", allow_module_level=True)
+# pytest.skip("skipping", allow_module_level=True)
 
 
-def fstring():
-    a = "1"
-    return f"{a}"
+def test_open_file(tmpdir):
+    import os
 
+    path = os.path.join(tmpdir, "text.111")
+    with open(path, "w") as f:
+        f.write("<file content>")
 
-def format_string():
-    a = "1"
-    return "%s" % (a)
-
-
-def test_string(benchmark):
-    result = benchmark(fstring)
-    assert result == "1"
-
-
-def test_format_string(benchmark):
-    result1 = benchmark(format_string)
-    assert result1 == "1"
+    assert os.path.isfile(path)